chore: move wasm binary to cmd/wasm
