+++ cmd/wasm/testdata/local_evaluation_inputs/invalid.json
